requirements.txt
requests==2.31.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.1
streamlit>=1.28.0
